fixed load/save properties currently it is profiling in release build changed to random shooting in order to do performance tests on xbox changed energy into a float-value (from int)
git-svn-id: https://svn.vis.ethz.ch/svn/kob_gamelab09/trunk@1656 fc16ba98-2866-4055-be8f-df951a5456ab

+++ src/ProjectMagma/ProjectMagma/Simulation/Properties/Environment/PowerUpControllerEnergy.cs
@@ -13,7 +13,7 @@ namespace ProjectMagma.Simulation
 
         protected override void GivePower(Entity player)
         {
-            player.SetInt("energy", player.GetInt("energy") + powerup.GetInt("value"));
+            player.SetFloat("energy", player.GetInt("energy") + powerup.GetInt("value"));
             Game.Instance.AudioPlayer.Play(Game.Instance.Simulation.SoundRegistry.PowerupEnergyTaken);
         }
